requirements.txt
streamlit>=1.40.0
pypdf>=4.0.0
python-docx>=1.1.0
langchain-text-splitters>=0.3.0
langchain-huggingface>=0.1.0
langchain-community>=0.3.0
sentence-transformers>=3.0.0
faiss-cpu>=1.7.4
langchain-groq>=0.2.0
pydantic>=2.9.0
python-dotenv>=1.0.1
